refactor(iframe-sandbox): the outer frame's script is a `.js` file imported as text
`outer-frame-script.js` is plain JavaScript under `// @ts-check`, so `deno
fmt`, `deno lint` and `deno check` all read it as code, and its comments need
no escaping. `outer-frame.ts` assembles the document around it: the CSP meta,
the styles, the inner frame, and one `<script>` element whose
`data-host-origin` attribute carries the host's origin, which the script reads
off its own element rather than from a global. Since the script's text is
inlined into that element, the module refuses at load time to carry a
`</script>`.
